requirements.txt
opencv-python
pillow
google.generativeai
easyocr
spacy
requests
tensorflow 
numpy
firebase-admin
datetime
scikit-image
argparse
geopy
mediapipe
face-recognition
langchain
langchain-google-genai
soundfile
torch
torchaudio
transformers
pyaudio
webrtcvad
rx
halo
onnx
onnxruntime
wheel
pyctcdecode
https://github.com/kpu/kenlm/archive/master.zip